requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_tflow_select=2.3.0=mkl
abseil-cpp=20211102.0=hd77b12b_0
absl-py=1.4.0=py310haa95532_0
aiofiles=22.1.0=py310haa95532_0
aiohttp=3.9.0=py310h2bbff1b_0
aiosignal=1.2.0=pyhd3eb1b0_0
altair=5.0.1=py310haa95532_0
annotated-types=0.6.0=py310haa95532_0
anyio=4.2.0=py310haa95532_0
arrow-cpp=11.0.0=h2c9b28c_2
astunparse=1.6.3=py_0
async-timeout=4.0.3=py310haa95532_0
attrs=23.1.0=py310haa95532_0
aws-c-common=0.4.57=ha925a31_1
aws-c-event-stream=0.1.6=hd77b12b_5
aws-checksums=0.1.9=ha925a31_0
aws-sdk-cpp=1.8.185=hd77b12b_0
beautifulsoup4=4.12.3=pyha770c72_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
blinker=1.6.2=py310haa95532_0
boost-cpp=1.82.0=h59b6b97_2
bottleneck=1.3.7=py310h9128911_0
brotli=1.0.9=h2bbff1b_7
brotli-bin=1.0.9=h2bbff1b_7
brotli-python=1.0.9=py310hd77b12b_7
bs4=4.12.3=hd8ed1ab_0
bzip2=1.0.8=he774522_0
c-ares=1.19.1=h2bbff1b_0
ca-certificates=2024.3.11=haa95532_0
cachetools=4.2.2=pyhd3eb1b0_0
certifi=2024.2.2=py310haa95532_0
cffi=1.16.0=py310h2bbff1b_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=py310haa95532_0
colorama=0.4.6=py310haa95532_0
contourpy=1.2.0=py310h59b6b97_0
cryptography=41.0.3=py310h3438e0d_0
cuda-cccl=12.4.99=0
cuda-cudart=11.8.89=0
cuda-cudart-dev=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-libraries-dev=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvrtc-dev=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-profiler-api=12.4.99=0
cuda-runtime=11.8.0=0
cycler=0.11.0=pyhd3eb1b0_0
datasets=2.12.0=py310haa95532_0
dill=0.3.6=py310haa95532_0
exceptiongroup=1.2.0=py310haa95532_0
fastapi=0.103.0=py310haa95532_0
ffmpeg=4.2.2=he774522_0
ffmpy=0.3.0=pyhb6f538c_0
filelock=3.13.1=py310haa95532_0
flatbuffers=2.0.0=h6c2663c_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=ha860e81_0
frozenlist=1.4.0=py310h2bbff1b_0
fsspec=2023.10.0=py310haa95532_0
fuzzywuzzy=0.18.0=py310haa95532_0
gast=0.4.0=pyhd3eb1b0_0
gflags=2.2.2=hd77b12b_1
giflib=5.2.1=h8cc25b3_3
glog=0.5.0=hd77b12b_1
gmpy2=2.1.2=py310h7f96b67_0
google-auth=2.22.0=py310haa95532_0
google-auth-oauthlib=0.4.4=pyhd3eb1b0_0
google-pasta=0.2.0=pyhd3eb1b0_0
gradio=4.21.0=pyhd8ed1ab_0
gradio-client=0.12.0=pyhd8ed1ab_0
grpc-cpp=1.48.2=hf108199_0
grpcio=1.42.0=py310hc60d5dd_0
h11=0.14.0=py310haa95532_0
h5py=3.9.0=py310hfc34f40_0
hdf5=1.12.1=h51c971a_3
httpcore=1.0.2=py310haa95532_0
httpx=0.26.0=py310haa95532_0
huggingface_hub=0.20.3=py310haa95532_0
icc_rt=2022.1.0=h6049295_2
icu=58.2=ha925a31_3
idna=3.4=py310haa95532_0
importlib-resources=6.1.1=pyhd3eb1b0_1
importlib_resources=6.1.1=py310haa95532_1
intel-openmp=2023.1.0=h59b6b97_46320
jinja2=3.1.3=py310haa95532_0
joblib=1.2.0=py310haa95532_0
jpeg=9e=h2bbff1b_1
jsonpickle=3.0.2=pypi_0
jsonschema=4.19.2=py310haa95532_0
jsonschema-specifications=2023.7.1=py310haa95532_0
kaggle=1.6.5=pyhd8ed1ab_0
keras=2.10.0=py310haa95532_0
keras-preprocessing=1.1.2=pyhd3eb1b0_0
kiwisolver=1.4.4=py310hd77b12b_0
lerc=3.0=hd77b12b_0
libboost=1.82.0=h3399ecb_2
libbrotlicommon=1.0.9=h2bbff1b_7
libbrotlidec=1.0.9=h2bbff1b_7
libbrotlienc=1.0.9=h2bbff1b_7
libcublas=11.11.3.6=0
libcublas-dev=11.11.3.6=0
libcufft=10.9.0.58=0
libcufft-dev=10.9.0.58=0
libcurand=10.3.5.119=0
libcurand-dev=10.3.5.119=0
libcurl=8.5.0=h86230a5_0
libcusolver=11.4.1.48=0
libcusolver-dev=11.4.1.48=0
libcusparse=11.7.5.86=0
libcusparse-dev=11.7.5.86=0
libdeflate=1.17=h2bbff1b_1
libevent=2.1.12=hcc03200_0
libffi=3.4.4=hd77b12b_0
libiconv=1.16=h2bbff1b_2
libjpeg-turbo=2.0.0=h196d8e1_0
libnpp=11.8.0.86=0
libnpp-dev=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libnvjpeg-dev=11.9.0.86=0
libpng=1.6.39=h8cc25b3_0
libprotobuf=3.20.3=h23ce68f_0
libssh2=1.10.0=hcd4344a_2
libthrift=0.15.0=he49ee6e_2
libtiff=4.5.1=hd77b12b_0
libuv=1.44.2=h2bbff1b_0
libwebp-base=1.3.2=h2bbff1b_0
libxml2=2.10.4=h0ad7f3c_1
libxslt=1.1.37=h2bbff1b_1
linkify-it-py=2.0.0=py310haa95532_0
lxml=4.9.3=py310h09808a7_0
lz4-c=1.9.4=h2bbff1b_0
markdown=3.4.1=py310haa95532_0
markdown-it-py=2.2.0=py310haa95532_1
markupsafe=2.1.3=py310h2bbff1b_0
matplotlib-base=3.8.0=py310h4ed8f06_0
mdit-py-plugins=0.3.0=py310haa95532_0
mdurl=0.1.0=py310haa95532_0
mkl=2023.1.0=h6b88ed4_46358
mkl-service=2.4.0=py310h2bbff1b_1
mkl_fft=1.3.8=py310h2bbff1b_0
mkl_random=1.2.4=py310h59b6b97_0
mpc=1.1.0=h7edee0f_1
mpfr=4.0.2=h62dcd97_1
mpir=3.0.0=hec2e145_1
mpmath=1.3.0=py310haa95532_0
multidict=6.0.4=py310h2bbff1b_0
multiprocess=0.70.14=py310haa95532_0
munkres=1.1.4=py_0
networkx=3.1=py310haa95532_0
nltk=3.8.1=py310haa95532_0
numexpr=2.8.7=py310h2cd9be0_0
numpy=1.26.3=py310h055cbcc_0
numpy-base=1.26.3=py310h65a83cf_0
oauthlib=3.2.2=py310haa95532_0
openjpeg=2.4.0=h4fc8c34_0
openssl=1.1.1w=h2bbff1b_0
opt_einsum=3.3.0=pyhd3eb1b0_1
orc=1.7.4=h623e30f_1
orjson=3.9.10=py310h2bbff1b_0
packaging=23.1=py310haa95532_0
pandas=2.2.1=py310h5da7b33_0
pillow=10.2.0=py310h2bbff1b_0
pip=23.3.1=py310haa95532_0
protobuf=3.20.3=py310hd77b12b_0
pyarrow=11.0.0=py310h790e06d_1
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=py_0
pycparser=2.21=pyhd3eb1b0_0
pydantic=2.5.3=py310haa95532_0
pydantic-core=2.14.6=py310h062c2fa_0
pydub=0.25.1=pyhd8ed1ab_0
pygments=2.15.1=py310haa95532_1
pyjwt=2.4.0=py310haa95532_0
pyopenssl=23.2.0=py310haa95532_0
pyparsing=3.0.9=py310haa95532_0
pysocks=1.7.1=py310haa95532_0
python=3.10.13=h966fe2a_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-flatbuffers=2.0=pyhd3eb1b0_0
python-levenshtein=0.12.2=py310h2bbff1b_0
python-multipart=0.0.9=pyhd8ed1ab_0
python-slugify=5.0.2=pyhd3eb1b0_0
python-tzdata=2023.3=pyhd3eb1b0_0
python-xxhash=2.0.2=py310h2bbff1b_1
python_abi=3.10=2_cp310
pytorch=2.2.1=py3.10_cuda11.8_cudnn8_0
pytorch-cuda=11.8=h24eeafa_5
pytorch-mutex=1.0=cuda
pytz=2023.3.post1=py310haa95532_0
pyyaml=6.0.1=py310h2bbff1b_0
rasa=3.6.16=pypi_0
re2=2022.04.01=hd77b12b_0
referencing=0.30.2=py310haa95532_0
regex=2023.10.3=py310h2bbff1b_0
requests=2.31.0=py310haa95532_0
requests-oauthlib=1.3.0=py_0
responses=0.13.3=pyhd3eb1b0_0
rich=13.3.5=py310haa95532_0
rpds-py=0.10.6=py310h062c2fa_0
rsa=4.7.2=pyhd3eb1b0_1
ruff=0.3.2=py310h298983d_0
safetensors=0.4.2=py310h87d50f1_0
scipy=1.11.4=py310h309d312_0
semantic_version=2.8.5=pyhd3eb1b0_0
setuptools=68.2.2=py310haa95532_0
shellingham=1.5.0=py310haa95532_0
six=1.16.0=pyhd3eb1b0_1
snappy=1.1.10=h6c2663c_1
sniffio=1.3.0=py310haa95532_0
soupsieve=2.5=py310haa95532_0
sqlite=3.41.2=h2bbff1b_0
starlette=0.27.0=py310haa95532_0
sympy=1.12=py310haa95532_0
tbb=2021.8.0=h59b6b97_0
tensorboard=2.10.0=py310haa95532_0
tensorboard-data-server=0.6.1=py310haa95532_0
tensorboard-plugin-wit=1.8.1=py310haa95532_0
tensorflow=2.10.0=mkl_py310hd99672f_0
tensorflow-base=2.10.0=mkl_py310h6a7f48e_0
tensorflow-estimator=2.10.0=py310haa95532_0
termcolor=2.1.0=py310haa95532_0
text-unidecode=1.3=pyhd3eb1b0_0
tk=8.6.12=h2bbff1b_0
tokenizers=0.15.1=py310h49fca51_0
tomlkit=0.12.0=pyha770c72_0
toolz=0.12.0=py310haa95532_0
torchaudio=2.2.1=pypi_0
torchvision=0.17.1=pypi_0
tqdm=4.65.0=py310h9909e9c_0
transformers=4.38.2=pyhd8ed1ab_0
typer=0.9.0=py310haa95532_0
typing-extensions=4.9.0=py310haa95532_1
typing_extensions=4.9.0=py310haa95532_1
tzdata=2023d=h04d1e81_0
uc-micro-py=1.0.1=py310haa95532_0
ucrt=10.0.20348.0=haa95532_0
ujson=5.4.0=py310hd77b12b_0
unidecode=1.2.0=pyhd3eb1b0_0
urllib3=1.26.18=py310haa95532_0
utf8proc=2.6.1=h2bbff1b_1
uvicorn=0.20.0=py310haa95532_0
vc=14.2=h21ff451_1
vc14_runtime=14.38.33130=h82b7239_18
vs2015_runtime=14.38.33130=hcb4865c_18
webencodings=0.5.1=py310haa95532_1
websockets=10.4=py310h2bbff1b_1
werkzeug=2.3.8=py310haa95532_0
wheel=0.41.2=py310haa95532_0
win_inet_pton=1.1.0=py310haa95532_0
wrapt=1.14.1=py310h2bbff1b_0
xxhash=0.8.0=h2bbff1b_3
xz=5.4.5=h8cc25b3_0
yaml=0.2.5=he774522_0
yarl=1.9.3=py310h2bbff1b_0
zlib=1.2.13=h8cc25b3_0
zstd=1.5.5=hd43e919_0
